requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
dj-database-url==1.2.0
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.14.0
flatbuffers==24.3.25
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.2
gunicorn==20.1.0
h5py==3.11.0
idna==3.7
joblib==1.4.0
keras==2.15.0
keras-tcn==3.5.0
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
ml-dtypes==0.3.2
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
platformdirs==4.2.1
protobuf==4.25.3
psycopg2-binary==2.9.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sqlparse==0.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-addons==0.22.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
typeguard==2.13.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.1
virtualenvwrapper-win==1.2.7
Werkzeug==3.0.2
whitenoise==6.2.0
wrapt==1.14.1